MET-867 Remove HarvesterClientImpll deactivation of jobs.
It was always failing with an exception. Now that that is not there the deactivation plugin will go through all the records and mark them deleted inside uim, this means a loading bar will be displayed. Before only the removal was done and then an exception was thrown so the deactivation was almost immediate.